requirements.txt
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.7
dnspython==2.2.1
Flask==2.0.3
Flask-Cors==3.0.10
Flask-PyMongo==0.5.2
gevent==23.9.1
greenlet==3.0.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
pymongo==3.12.3
python-dateutil==2.8.2
requests==2.27.0
six==1.16.0
tornado==6.1
urllib3==1.26.18
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.1
